+++ tests/Test/ConfigMysqlCacheDbTest.php
@@ -62,7 +62,7 @@ class ConfigMysqlCacheDbTest extends ConfigMysqlTest
      */
     public function testLoadFixturesAndCheckBackup(): void
     {
-        $this->loadFixtures([
+        $this->databaseTool->loadFixtures([
             'Liip\Acme\Tests\App\DataFixtures\ORM\LoadUserData',
         ]);
 
@@ -95,7 +95,7 @@ class ConfigMysqlCacheDbTest extends ConfigMysqlTest
         $salt = $user1->getSalt();
 
         // Clean database
-        $this->loadFixtures();
+        $this->databaseTool->loadFixtures();
 
         $users = $this->entityManager->getRepository('LiipAcme:User')
             ->findAll();
@@ -107,7 +107,7 @@ class ConfigMysqlCacheDbTest extends ConfigMysqlTest
         );
 
         // Load fixtures again
-        $this->loadFixtures([
+        $this->databaseTool->loadFixtures([
             'Liip\Acme\Tests\App\DataFixtures\ORM\LoadUserData',
         ]);
 
@@ -138,13 +138,13 @@ class ConfigMysqlCacheDbTest extends ConfigMysqlTest
      */
     public function testLoadFixturesCheckReferences(): void
     {
-        $referenceRepository = $this->loadFixtures([
+        $referenceRepository = $this->databaseTool->loadFixtures([
             'Liip\Acme\Tests\App\DataFixtures\ORM\LoadUserData',
         ])->getReferenceRepository();
 
         $this->assertCount(1, $referenceRepository->getReferences());
 
-        $referenceRepository = $this->loadFixtures([
+        $referenceRepository = $this->databaseTool->loadFixtures([
             'Liip\Acme\Tests\App\DataFixtures\ORM\LoadUserData',
             'Liip\Acme\Tests\App\DataFixtures\ORM\LoadSecondUserData',
         ])->getReferenceRepository();